Rename save_to_model(s) methods to sync_to_model(s). Close #3
