+++ spec/mappable/sync_to_models/simple_form_few_models_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
+RSpec.describe 'sync_to_models: Simple Form Object - Few Models' do
   let(:team) { Struct.new(:team_name, :year).new }
   let(:car) { {} }
 
@@ -10,21 +10,21 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
 
   context 'without default model' do
     module SaveToModels
-      class SimpleFormName < FormObj::Form
+      class SimpleForm < FormObj::Form
         Engine = Struct.new(:power)
 
         include FormObj::ModelMapper
 
         attribute :name, model_attribute: :team_name, model: :team
         attribute :year, model: :team
-        attribute :engine_power, model: :car, model_attribute: ':engine.power', model_class: 'SaveToModels::SimpleFormName::Engine'
+        attribute :engine_power, model: :car, model_attribute: ':engine.power', model_class: Engine
       end
     end
-    let(:form) { SaveToModels::SimpleFormName.new }
+    let(:form) { SaveToModels::SimpleForm.new }
 
     context 'nested models are created when they do not exist yet' do
       it 'has all attributes correctly saved' do
-        form.save_to_models(team: team, car: car)
+        form.sync_to_models(team: team, car: car)
 
         expect(team.team_name).to     eq form.name
         expect(team.year).to          eq form.year
@@ -32,7 +32,7 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
       end
 
       it 'returns self' do
-        expect(form.save_to_models(team: team, car: car)).to eql form
+        expect(form.sync_to_models(team: team, car: car)).to eql form
       end
     end
 
@@ -40,7 +40,7 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
       let(:car) {{ engine: Struct.new(:power).new }}
 
       it 'has all attributes correctly saved' do
-        form.save_to_models(team: team, car: car)
+        form.sync_to_models(team: team, car: car)
 
         expect(team.team_name).to     eq form.name
         expect(team.year).to          eq form.year
@@ -48,28 +48,28 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
       end
 
       it 'returns self' do
-        expect(form.save_to_models(team: team, car: car)).to eql form
+        expect(form.sync_to_models(team: team, car: car)).to eql form
       end
     end
   end
 
   context 'with default model' do
     module SaveToModelsWithDefault
-      class SimpleFormName < FormObj::Form
+      class SimpleForm < FormObj::Form
         Engine = Struct.new(:power)
 
         include FormObj::ModelMapper
 
         attribute :name, model_attribute: :team_name
         attribute :year
-        attribute :engine_power, model: :car, model_attribute: ':engine.power', model_class: 'SaveToModelsWithDefault::SimpleFormName::Engine'
+        attribute :engine_power, model: :car, model_attribute: ':engine.power', model_class: Engine
       end
     end
-    let(:form) { SaveToModelsWithDefault::SimpleFormName.new }
+    let(:form) { SaveToModelsWithDefault::SimpleForm.new }
 
     context 'nested models are created when they do not exist yet' do
       it 'has all attributes correctly saved' do
-        form.save_to_models(default: team, car: car)
+        form.sync_to_models(default: team, car: car)
 
         expect(team.team_name).to     eq form.name
         expect(team.year).to          eq form.year
@@ -77,7 +77,7 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
       end
 
       it 'returns self' do
-        expect(form.save_to_models(default: team, car: car)).to eql form
+        expect(form.sync_to_models(default: team, car: car)).to eql form
       end
     end
 
@@ -85,7 +85,7 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
       let(:car) {{ engine: Struct.new(:power).new }}
 
       it 'has all attributes correctly saved' do
-        form.save_to_models(default: team, car: car)
+        form.sync_to_models(default: team, car: car)
 
         expect(team.team_name).to     eq form.name
         expect(team.year).to          eq form.year
@@ -93,7 +93,7 @@ RSpec.describe 'save_to_models: Simple Form Object - Few Models - Name' do
       end
 
       it 'returns self' do
-        expect(form.save_to_models(default: team, car: car)).to eql form
+        expect(form.sync_to_models(default: team, car: car)).to eql form
       end
     end
   end
