+++ spec/mappable/sync_to_models/simple_form_one_model_spec.rb
@@ -1,26 +1,28 @@
-RSpec.describe 'save_to_model: Simple Form Object - One Model - Implicit Model Class' do
-  Engine = Struct.new(:power)
-
+RSpec.describe 'sync_to_model: Simple Form Object - One Model' do
   module SaveToModel
-    module ImplicitModelClass
-      class SimpleForm < FormObj::Form
-        include FormObj::ModelMapper
+    class SimpleForm < FormObj::Form
+      Engine = Struct.new(:power)
 
-        attribute :name, model_attribute: :team_name
-        attribute :year
-        attribute :engine_power, model_attribute: 'car.:engine.power'
-      end
+      include FormObj::ModelMapper
+
+      attribute :name, model_attribute: :team_name
+      attribute :year
+      attribute :month, model_attribute: false
+      attribute :day, model_attribute: false
+      attribute :engine_power, model_attribute: 'car.:engine.power', model_class: [Hash, Engine]
     end
   end
 
-  let(:engine) { Engine.new }
-  let(:model) { Struct.new(:team_name, :year, :car).new }
-  let(:form) { SaveToModel::ImplicitModelClass::SimpleForm.new() }
+  let(:engine) { SaveToModel::SimpleForm::Engine.new }
+  let(:model) { Struct.new(:team_name, :year, :month, :car).new }
+  let(:form) { SaveToModel::SimpleForm.new() }
 
   shared_context 'fill in a form' do
     before do
       form.name = 'Ferrari'
       form.year = 1950
+      form.month = 'April'
+      form.day = 1               # will not raise error if attribute is not existent in the model
       form.engine_power = 335
     end
   end
@@ -29,13 +31,14 @@ RSpec.describe 'save_to_model: Simple Form Object - One Model - Implicit Model C
     it 'creates non-existent models and correctly saves all attributes' do
       expect(model.team_name).to          eq form.name
       expect(model.year).to               eq form.year
+      expect(model.month).to              be_nil
       expect(model.car[:engine].power).to eq form.engine_power
     end
   end
 
   context 'nested models do not exists' do
     include_context 'fill in a form'
-    before { form.save_to_model(model) }
+    before { form.sync_to_model(model) }
 
     it_behaves_like 'a form that can be saved'
   end
@@ -45,7 +48,7 @@ RSpec.describe 'save_to_model: Simple Form Object - One Model - Implicit Model C
     before { model.car = car }
 
     include_context 'fill in a form'
-    before { form.save_to_model(model) }
+    before { form.sync_to_model(model) }
 
     it_behaves_like 'a form that can be saved'
     it "doesn't create existing nested models" do
@@ -58,7 +61,7 @@ RSpec.describe 'save_to_model: Simple Form Object - One Model - Implicit Model C
     before { model.car = car }
 
     include_context 'fill in a form'
-    before { form.save_to_model(model) }
+    before { form.sync_to_model(model) }
 
     it_behaves_like 'a form that can be saved'
     it "doesn't create existing nested models" do
